syncval: Minimize include of sync_validation.h
Needed to allow queue/global operations to breakout cleanly
